+++ 00-demo/00-demo.js
@@ -1,5 +1,6 @@
 const express = require('express')
 const expressHandlebars = require('express-handlebars')
+const fortune = require('./lib/fortune.js')
 const app = express()
 
 app.engine('handlebars', expressHandlebars.engine({
@@ -14,17 +15,10 @@ const port = process.env.PORT || 3000
 app.get('/', (req, res) => res.render('home'))
 
 
-const fortunes = [
-    "Conquer your fears or they will conquer you.", "Rivers need springs.",
-    "Do not fear what you don't know.",
-    "You will have a pleasant surprise.",
-    "Whenever possible, keep it simple.",
-]
+
 app.get('/about', (req, res) => {
-    const randomFortune =
-    fortunes[Math.floor(Math.random()*fortunes.length)]
-    res.render('about', { fortune: randomFortune })
-    })
+    res.render('about', { fortune: fortune.getFortune() })
+})
 app.use((req, res) => {
     res.status(404)
     res.render('404')
